Move functions out of PointSearch class
